% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cap.R
\name{add_cap}
\alias{add_cap}
\title{Cap a Numeric Variable}
\usage{
add_cap(dataset, var_qt, cap_value, cap_color, type = "filter")
}
\arguments{
\item{dataset}{Data set for which to apply the cap to}

\item{var_qt}{Character; the variable of interest (in quotation marks) for which to apply the cap}

\item{cap_value}{Numeric; the value of the specified variable (\code{var_qt}) at and above to be capped}

\item{cap_color}{Character; the color which will be applied to values at or above \code{cap_value}}

\item{type}{Character choice: "filter" or "flag"; the former will replace values at or above \code{cap_value} with 'NA', while the latter will add a logical column ("above_cap") for if the respective row's value is at or above the \code{cap_value}.}
}
\value{
Character versions rounded to the number of digits specified, including trailing zeroes
}
\description{
Apply a cap value of a specified color to numeric variables, either by replacing them with 'NA', or flagging them in a new column.
}
\examples{
cap_results <- add_cap(data_pm25, "pm25_atm", 150, "red")
ggplot() + cap_results$cap_guide + labs(subtitle = cap_results$lab_subtitle_cap)
}
