% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_date_tags.R
\name{apply_date_tags}
\alias{apply_date_tags}
\title{Create a column of "Date Tags"}
\usage{
apply_date_tags(
  dataset,
  starts = input_date_starts,
  ends = input_date_ends,
  tags = input_date_tags,
  stamp = "17 Jan 1999"
)
}
\arguments{
\item{dataset}{The dataset for which to apply the date tags to.}

\item{starts}{Character list of start dates (format for each date: "YYYY-MM-DD")}

\item{ends}{Character list of end dates (format for each date: "YYYY-MM-DD")}

\item{tags}{Character list of tags}

\item{stamp}{Desired date stamp format type. To avoid ambiguity, specify a day example with a value greater than 12, and a 1990s year.}
}
\value{
Dataframe with a new column ("date_tag") containing the appropriate date tag with respect to each row's time stamp.
}
\description{
Create a column of date tags based on specified date ranges (\code{starts} and \code{ends}) and respective names (\code{tags}).
}
\examples{
  apply_date_tags(
    july_api_daily \%>\% filter(site_id == "0441c6b9f431d0e4_23361") \%>\% select(site_id, date, temperature),
    starts = c("2020-07-01", "2020-07-04", "2020-07-05"),
    ends = c("2020-07-03", "2020-07-04", "2020-07-07"),
    tags = c("Before", "Independence Day", "After")
  )
}
