% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_epa.R
\name{apply_epa}
\alias{apply_epa}
\title{Apply EPA Corrections}
\usage{
apply_epa(
  dataset,
  by_day = TRUE,
  by_hour = FALSE,
  epa_percent = 75,
  keep_cols = FALSE
)
}
\arguments{
\item{dataset}{The dataset for which to apply the correction factors to}

\item{by_day}{Logical; average data by day}

\item{by_hour}{Logical; average data by hour}

\item{epa_percent}{Numeric; Minimum percentage of data required to be included}

\item{keep_cols}{Logical; Keep or disgard extra columns. If FALSE, only identifying columns and EPA-corrected columns will remain}
}
\value{
Dataframe with new columns for EPA-corrected PM2.5, and values removed if A & B determined to be in disagreement or minimum data requirement not met.
}
\description{
Apply EPA correction factors to PurpleAir PM2.5 data.
The EPA has established conditions in 2020 and late 2020 (the latter of which is tagged as "2021" in this set) for correcting PurpleAir data, including minimum data requirements and checks for A & B monitor agreement.
The late 2020 correction factor includes consideration for high-PM events (such as wildfires); the low-PM values have the same correction factor as 2020.
Unlike the LRAPA correction factor (\link{apply_lrapa}) which uses CF=ATM PM2.5 values, the major EPA correction factors use CF=1 values. If needed, an additional EPA-based output column using CF=ATM is included. More information on this variable can be found on the \href{https://www2.purpleair.com/community/faq#hc-what-is-the-difference-between-cf-1-and-cf-atm}{PurpleAir FAQ}.
More information about EPA correction factors can be found on the \href{https://www.epa.gov/air-sensor-toolbox/technical-approaches-sensor-data-airnow-fire-and-smoke-map}{Air Sensor Toolbox}.
}
\examples{
data_with_epa <- apply_epa(raw_data, keep_cols = TRUE)
epa_hourly <- apply_epa(raw_data, by_hour = TRUE)
}
