% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_functions.R
\name{apply_functions}
\alias{apply_functions}
\title{Apply Corrections & Tags to Dataset}
\usage{
apply_functions(
  dataset,
  by_day = TRUE,
  by_hour = FALSE,
  tag_dates = run_date_grouping,
  tag_hours = run_hour_grouping
)
}
\arguments{
\item{dataset}{The dataset for which to apply the date tags to.}

\item{by_day}{Logical; Average by day}

\item{by_hour}{Logical; Average by hour}

\item{tag_dates}{Logical; Apply date tags}

\item{tag_hours}{Logical; Apply hour tags}
}
\value{
Dataframe with new columns for corrected values and appropriate tagged values ("date_tag" and "hour_tag") containing the appropriate hour/date tag with respect to each row's timestamp.
}
\description{
Average a dataset by specified time stamp groupings, apply correction factors (using \link{apply_corrections}, which uses \link{apply_epa} and \link{apply_lrapa}), and tag dates/hours (\link{apply_date_tags} and \link{apply_hour_tags}, respectively) as appropriate.
See \link{apply_date_tags} and \link{apply_hour_tags} for more information individually.
}
\examples{
daily_data <- apply_functions(raw_data, by_day = TRUE, by_hour = TRUE)
}
