% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_hour_tags.R
\name{apply_hour_tags}
\alias{apply_hour_tags}
\title{Create a column of "Date Tags"}
\usage{
apply_hour_tags(dataset, starts = input_hour_starts, tags = input_hour_tags)
}
\arguments{
\item{dataset}{The dataset for which to apply the date tags to.}

\item{starts}{Numeric list of start hours (in 24 hour format)}

\item{tags}{Character list of tags}
}
\value{
Dataframe with a new column ("hour_tag") containing the appropriate hour tag with respect to each row's timestamp.
}
\description{
Create a column of date tags based on specified hour ranges (created using inputted \code{starts}) and respective names (\code{tags}).
}
\examples{
tagged_data <- apply_hour_tags(data_hourly, starts = c(5, 12, 17, 21), tags = c("Morning", "Afternoon", "Evening", "Night"))
}
