% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_lrapa.R
\name{apply_lrapa}
\alias{apply_lrapa}
\title{Apply LRAPA Corrections}
\usage{
apply_lrapa(dataset, by_day = TRUE, by_hour = FALSE, keep_cols = FALSE)
}
\arguments{
\item{dataset}{The dataset for which to apply the correction factors to}

\item{by_day}{Logical; average data by day}

\item{by_hour}{Logical; average data by hour}

\item{keep_cols}{Logical; Keep or disgard extra columns. If FALSE, only identifying columns and LRAPA-corrected columns will remain}
}
\value{
Dataframe with new columns for LRAPA-corrected PM2.5.
}
\description{
Apply LRAPA correction factors to PurpleAir PM2.5 data.
\href{https://www.lrapa.org/}{LRAPA} established PM2.5 correction factors for PurpleAir's PM2.5 CF=ATM values.
Unlike the EPA correction factors (\link{apply_epa}) which uses CF=1 PM2.5 values, the LRAPA correction factor uses CF=ATM values. More information on this variable can be found on the \href{https://www2.purpleair.com/community/faq#hc-what-is-the-difference-between-cf-1-and-cf-atm}{PurpleAir FAQ}.
It should also be noted that this correction factor might not apply to all airsheds, and was developed in Oregon, where (along with much of the Pacific Northwest) woodsmoke is a major contributor to PM2.5.
For more information on the correction factor, see the \href{https://www.lrapa.org/DocumentCenter/View/4147/PurpleAir-Correction-Summary}{LRAPA documentation}.
}
\examples{
data_with_lrapa <- apply_lrapa(raw_data, keep_cols = TRUE)
lrapa_hourly <- apply_lrapa(raw_data, by_hour = TRUE)
}
