% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius conversion}
\usage{
C_to_F(C_temp)
}
\arguments{
\item{C_temp}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert degrees Celsius temperatures to degrees Fahrenheit
}
\examples{
temp1 <- C_to_F(22);
temp2 <- C_to_F( c(-2, 12, 23) );
}
