% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_incomplete.R
\name{drop_incomplete}
\alias{drop_incomplete}
\title{Drop Incomplete Sets}
\usage{
drop_incomplete(dataset, var_qt)
}
\arguments{
\item{dataset}{Data set for which to filter incomplete monitor data}

\item{var_qt}{Character; the variable of interest (in quotation marks) for which to apply the cap}
}
\value{
Data set with monitors without complete sets removed
}
\description{
Drop monitor data for which its time series data is incomplete compared to that of other monitors in the data set.
}
\examples{
drop_incomplete(july_api_daily, "pm25_atm")
}
\seealso{
\code{\link{heatmap_single()}} and \code{\link{heatmap_cross()}}

Other {visualization settings}: 
\code{\link{add_cap}()},
\code{\link{settings_dt_scale}()},
\code{\link{settings_shapes}()},
\code{\link{settings_units}()}
}
\concept{{visualization settings}}
