% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_df.R
\name{filter_df}
\alias{filter_df}
\title{Filter a Data Frame}
\usage{
filter_df(
  dataframe,
  var = label,
  include = c(""),
  exclude = c(""),
  location_data = data_meta
)
}
\arguments{
\item{var}{The varaible which to filter (not in quotation marks)}

\item{include}{Character list; the list of values to keep for the data frame (optional if \code{exclude} specified)}

\item{exclude}{Character list; the list of values to remove from the data frame (optional if \code{include} specified)}

\item{location_data}{Optional, unless \code{var} not in \code{dataset}; the data set containing monitor meta data}

\item{dataset}{The dataset which to filter}
}
\value{
Data frame filtered by the specifications above
}
\description{
Filter a data frame based on any character variable, including varaibles from meta data, without having to conduct joins or string arguments yourself.
}
\examples{
filter_df(july_api_diurnal, hour_tag, include = c("Morning", "Afternoon"))
filter_df(july_api_diurnal, label, exclude = c("Lighthouse"), location_data = july_api_meta)
}
\seealso{
Other {miscellaneous functions}: 
\code{\link{adjust_timezone}()},
\code{\link{ambient_temperature}()},
\code{\link{apply_date_tags}()},
\code{\link{apply_hour_tags}()},
\code{\link{column_dt}()},
\code{\link{group_stad}()},
\code{\link{organize_stad}()},
\code{\link{rounding_w_zeroes}()},
\code{\link{unit_convert}()},
\code{\link{wrangle_meta}()}
}
\concept{{miscellaneous functions}}
