% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_pas.R
\name{get_area_pas}
\alias{get_area_pas}
\title{Getting PAS by Area}
\usage{
get_area_pas(
  state_code = input_stateCode,
  west = input_west,
  east = input_east,
  south = input_south,
  north = input_north,
  labels = input_labels,
  datestamp = input_enddate,
  startdate = input_startdate
)
}
\arguments{
\item{state_code}{Optional. State of monitors to include. If NULL/unspecified, all monitors specified in the above boundaries will be included regardless of state.}

\item{west}{The western boundary coordinate}

\item{east}{The eastern boundary coordinate}

\item{south}{The southern boundary coordinate}

\item{north}{The northern boundary coordinate}

\item{labels}{Labels of monitors to include. If NULL/unspecified, all monitors will be included. The labels are based on string detection, so for instance having "STAR" will pull all that have the word "STAR" in the label. Capitalization matters.}

\item{datestamp}{Important argument for loading historical data, as not all monitors might be actively reporting anymore}

\item{startdate}{Relevant for calculating lookback days (the maximum number of days to go back and try to load data if the requested date cannot be retrieved)}
}
\value{
A dataframe returning all PAS within the defined area
}
\description{
Load the PAS within a specified area
}
\examples{
area_pas_defaults <- get_area_pas()
area_pas_custom <- get_area_pas("OR", -122.854, 45.4, -122.58, 45.6, c("se", "SE", "Se", "\\\\bSTAR\\\\b", "\\\\bPSU\\\\b"), datestamp = "2021-07-07", startdate = "2021-07-01")
}
