% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_pat.R
\name{get_area_pat}
\alias{get_area_pat}
\title{Get PurpleAir Timeseries (PAT) raw data and sensor meta data}
\usage{
get_area_pat(
  id_list = ids,
  pas_input = pas_area,
  startdate = input_startdate,
  enddate = input_enddate,
  cols = c("created_at", "temperature", "humidity", "pm2.5_cf1", "pm2.5_atm")
)
}
\arguments{
\item{id_list}{Character list; A list of PAS IDs}

\item{pas_input}{PAS object; see \link{get_area_pas}.}

\item{startdate}{Character; Date at which to start the collection (format: "YYYY-MM-DD")}

\item{enddate}{Character; Date through which the data will be collected (format: "YYYY-MM-DD")}

\item{cols}{Character list; List of variables of interest to keep. Site ID info will autmatically be collected}
}
\value{
List: dataframes of sensor meta info, and a dataframe of raw data for the variables of interest
}
\description{
Iterate over a list of PurpleAir sensor IDs to extract sensor meta data and timeseries data given specified start and end dates
}
\examples{
\donttest{
get_area_pat(
  id_list = c("6c18a7181a09037c_47107", "9f6643766cd4ba03_7572"),
  pas_input = (example(get_area_pas))$value,
  startdate = "2020-07-01", enddate = "2020-07-03"
)
}
}
\seealso{
\code{\link[AirSensor]{pat_downloadParseRawData()}}

Other {PA API functions}: 
\code{\link{get_area_pas}()},
\code{\link{get_ids}()},
\code{\link{wrangle_data}()}
}
\concept{{PA API functions}}
