% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_ids}
\alias{get_ids}
\title{Get IDs from a PAS}
\usage{
get_ids(pas = pas_area, outside = include_outside, inside = include_inside)
}
\arguments{
\item{pas}{A dataframe of PAS containing a column of device deployment IDs}

\item{outside}{Logical argument to include/exclude outdoor monitors}

\item{inside}{Logical argument to include/exclude indoor monitors}
}
\value{
A list of IDs from a provided PAS
}
\description{
Load a list of device deployment IDs from a provided PAS
}
\examples{
id_default <- get_ids()
id_outside <- get_ids(outside = TRUE, inside = FALSE)
}
