% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{group_data}
\alias{group_data}
\title{Average Data by Time Groupings}
\usage{
group_data(dataset, by_day = TRUE, by_hour = FALSE)
}
\arguments{
\item{dataset}{The dataset for which to average}

\item{by_day}{Logical; average data by day}

\item{by_hour}{Logical; average data by hour}
}
\value{
Dataframe with numeric values averaged by specified timestamp(s)
}
\description{
Average numeric values in a dataset by day and/or by hour.
}
\examples{
data_daily <- group_data(raw_data, by_day = TRUE)
data_hourly <- group_data(raw_data, by_day = TRUE, by_hour = FALSE)
}
