% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_stad.R
\name{group_stad}
\alias{group_stad}
\title{Group temporal data by a specified time unit}
\usage{
group_stad(dataset, by_day = TRUE, by_hour = FALSE)
}
\arguments{
\item{dataset}{The data set for which to average}

\item{by_day}{Logical; average data by day}

\item{by_hour}{Logical; average data by hour}
}
\value{
Data frame with numeric values averaged by specified time stamp(s)
}
\description{
Average numeric values in a data set by day and/or by hour. This is intended to be applied to raw or otherwise high-frequency temporal data.
}
\examples{
group_stad(july_api_full)
\donttest{
group_stad(july_api_full, by_day = FALSE, by_hour = TRUE)
group_stad(july_api_full, by_day = TRUE, by_hour = TRUE)
}
}
\seealso{
\code{\link{column_dt()}}

Other {miscellaneous functions}: 
\code{\link{adjust_timezone}()},
\code{\link{ambient_temperature}()},
\code{\link{apply_date_tags}()},
\code{\link{apply_hour_tags}()},
\code{\link{column_dt}()},
\code{\link{filter_df}()},
\code{\link{organize_stad}()},
\code{\link{rounding_w_zeroes}()},
\code{\link{unit_convert}()},
\code{\link{wrangle_meta}()}
}
\concept{{miscellaneous functions}}
