% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_cross.R
\name{heatmap_cross}
\alias{heatmap_cross}
\title{Heatmap for multiple monitors' temporal atmospheric data}
\usage{
heatmap_cross(
  dataset,
  variable_of_interest,
  drop_incomplete = FALSE,
  cap_value = NA,
  cap_color = "red",
  data_labels = FALSE,
  text_color = "black",
  location_data = data_meta,
  digits = 2
)
}
\arguments{
\item{dataset}{The hourly dataset to visualize}

\item{variable_of_interest}{The variable of interest (not in quotation marks) which to visualize}

\item{drop_incomplete}{Logical; drop/keep incomplete monitors, see \link{drop_incomplete} for more information.}

\item{cap_value}{Numeric, optional; values at or above to be colored serpately from the regular continuous scale. See \link{add_cap} for more information.}

\item{cap_color}{Character; color for values at or above the \code{cap_value}}

\item{data_labels}{Logical; label each cell in the heatmap with the appropriate value}

\item{text_color}{Character; the color of data label text}

\item{location_data}{Data set containing latitude and longitude data}

\item{digits}{Numeric; the number of digits to report}
}
\value{
Data visualization: heatmap colored by a specified numeric variable, with time on the x-axis (and appropriate breaks & labels) and monitor labels on the y-axis arranged from north to south and separted by location.
}
\description{
Visualize temporal atmospheric data for multiple monitors using a heatmap, with optional data labels. To visualize one monitor on a heatmap, see \link{heatmap_single}.
Relevant information (such as date ranges, averaging methods, facets, and min/max values in the set) will be reported automatically in the visualization.
}
\examples{
\donttest{
heatmap_cross(july_api_hourly, pm25_atm, location_data = july_api_meta)
heatmap_cross(july_api_hourly, pm25_atm, location_data = july_api_meta, drop_incomplete = TRUE, cap_value = 75, cap_color = "green")
heatmap_cross(july_api_daily, temperature, location_data = july_api_meta, drop_incomplete = TRUE)
}
}
\seealso{
Other {STAD visualizations}: 
\code{\link{heatmap_single}()},
\code{\link{map_stad}()},
\code{\link{ts_line}()},
\code{\link{ts_variation}()}
}
\concept{{STAD visualizations}}
