% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_single.R
\name{heatmap_single}
\alias{heatmap_single}
\title{Heatmap for a single monitor's temporal atmospheric data}
\usage{
heatmap_single(
  variable_of_interest,
  site_of_interest = "",
  cap_value = NA,
  cap_color = "red",
  data_labels = TRUE,
  digits = 2,
  date_breaks = "1 day",
  text_color = "black",
  dataset = data_hourly,
  location_data = data_meta
)
}
\arguments{
\item{variable_of_interest}{The variable of interest (not in quotation marks) which to visualize}

\item{site_of_interest}{Character; the label (or a portion of the label) of the monitor to visualize}

\item{cap_value}{Numeric, optional; values at or above to be colored serpately from the regular continuous scale. See \link{add_cap} for more information.}

\item{cap_color}{Character; color for values at or above the \code{cap_value}}

\item{data_labels}{Logical; label each cell in the heatmap with the appropriate value}

\item{digits}{Numeric; the number of digits to report}

\item{date_breaks}{Character; the frequency of x-axis label breaks}

\item{text_color}{Character; the color of data label text}

\item{dataset}{The hourly data set to visualize}

\item{location_data}{Data set containing latitude and longitude data}
}
\value{
Data visualization: hourly heatmap colored by a specified numeric variable, with date on the x-axis and hours on the y-axis.
}
\description{
Visualize hourly atmospheric data for a single monitor using a heatmap, with optional data labels. To visualize multiple monitors on a heatmap, see \link{heatmap_cross}.
Relevant information (such as date ranges, averaging methods, facets, and min/max values in the set) will be reported automatically in the visualization.
}
\examples{
\donttest{
heatmap_single(pm25_epa_2021, "Lighthouse", dataset = july_api_hourly, location_data = july_api_meta)
heatmap_single(temperature, "Lighthouse", cap_value = 85, cap_color = "green", data_label = FALSE, dataset = july_api_hourly, location_data = july_api_meta)
}
}
\seealso{
Other {STAD visualizations}: 
\code{\link{heatmap_cross}()},
\code{\link{map_stad}()},
\code{\link{ts_line}()},
\code{\link{ts_variation}()}
}
\concept{{STAD visualizations}}
