% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_stad.R
\name{organize_stad}
\alias{organize_stad}
\title{Combine and/or Pivot STAD Data Set(s)}
\usage{
organize_stad(
  dataset,
  measurements,
  col_rename,
  keep_cols = FALSE,
  drop_missing = TRUE
)
}
\arguments{
\item{dataset}{Dataset(s) to pivot. If more than one data set is provided, they will be row-bound. Use \code{list()} if more than one.}

\item{measurements}{Character, optional. Column headers to pivot. Use \code{c()}.}

\item{col_rename}{Character, optional, only run if \code{measurements} provided. The character to rename the 'value' column output of the pivot to.}

\item{keep_cols}{Logical, only run if \code{measurements} provided. Keep (\code{TRUE}) or disgard (\code{FALSE}) the non-identifying columns that have not been pivoted.}

\item{drop_missing}{Logical, only run if \code{measurements} provided. Drop missing values from pivoted column. Recommended especially if \code{keep_cols} is \code{FALSE}.}
}
\value{
Data set.
}
\description{
Choose to combine multiple STAD data sets, and/or pivot columns. Useful for combining FRM and PA data frames, or pivoting columns of the same pollutant type.
}
\examples{
# Pivot columns of a similar pollutant:
head(organize_stad(july_api_daily, c(pm25_epa_2021, pm25_epa_2020, pm25_lrapa), "pm25"))
# Combine PA and FRM data:
head(organize_stad(list(july_api_daily, july_frm_daily), c(pm25_epa_2021, pm25_frm)))
# Combine meta data:
tail(organize_stad(list(july_api_meta, july_frm_meta)))
}
