% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_frm.R
\name{read_frm}
\alias{read_frm}
\title{Wrangle federal reference monitor (FRM)}
\usage{
read_frm(file_path = input_frm_path)
}
\arguments{
\item{file_path}{File path to the Excel sheet of FRM data set to wrangle.}
}
\value{
A raw R data set, prepared to be passed to \link{wrangle_frm}.
}
\description{
The following is based on inputs assuming an Excel input from \href{https://oraqi.deq.state.or.us/report/SingleStationReport}{Oregon DEQ single station report}.
}
\examples{
\donttest{
2+2
}
}
\seealso{
Other {FRM functions}: 
\code{\link{wrangle_frm}()}
}
\concept{{FRM functions}}
