% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_local_folder.R
\name{read_local_folder}
\alias{read_local_folder}
\title{Read a Local Folder of PurpleAir Data}
\usage{
read_local_folder(
  folder_path = input_path,
  read_primary = TRUE,
  read_secondary = TRUE,
  pattern = NULL
)
}
\arguments{
\item{folder_path}{Character; the file path for the folder that contains the monitor data frames}

\item{read_primary}{Logical; read primary data}

\item{read_secondary}{Logical; read secondary data}

\item{pattern}{Character, optional; filter for a specific location (by label in the file title)}
}
\value{
Raw R data frame
}
\description{
Read a folder containing PurpleAir CSV data.
}
\examples{
\dontrun{
read_local_folder(folder_path = "data-raw/local")
read_local_folder(folder_path = "data-raw/local", pattern = "Lincoln")
}
}
\seealso{
\code{\link{read_local_single()}} to read single files in a folder.

Other {PA local functions}: 
\code{\link{read_local_meta}()},
\code{\link{read_local_single}()},
\code{\link{wrangle_local}()}
}
\concept{{PA local functions}}
