% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_local_meta.R
\name{read_local_meta}
\alias{read_local_meta}
\title{Create a data frame of locally loaded monitor data}
\usage{
read_local_meta(
  folder_path = input_path,
  pattern = NULL,
  timezone = input_timezone
)
}
\arguments{
\item{folder_path}{Character; the file path for the folder that contains the monitor data frames}

\item{pattern}{Character, optional; filter for a specific location (by label in the file title)}

\item{timezone}{Character of valid Olson name for timezone of the monitors}
}
\value{
A wrangled data frame of monitor location data from a local folder of PurpleAir data
}
\description{
Load PurpleAir monitor data based on locally stored data, extracting relevant information from the file names. See
}
\examples{
\dontrun{
read_local_meta("data-raw/local", timezone = "America/Los_Angeles")
read_local_meta("data-raw/local", timezone = "America/Los_Angeles", pattern = "Lincoln")
}
}
\seealso{
Other {PA local functions}: 
\code{\link{read_local_folder}()},
\code{\link{read_local_single}()},
\code{\link{wrangle_local}()}
}
\concept{{PA local functions}}
