% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_local_single.R
\name{read_local_single}
\alias{read_local_single}
\title{Title}
\usage{
read_local_single(
  input_file,
  selects = c("pm25_cf1", "pm25_atm", "temperature", "humidity")
)
}
\arguments{
\item{input_file}{Character; a file path}

\item{selects}{Character list of the renamed variables to keep. Time stamp automatically kept.}
}
\value{
Data frame of input file with relevant info in columns
}
\description{
Read specific file by location, skipping columns as needed; extracting info from title
}
