% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oa_folder.R
\name{read_oa_folder}
\alias{read_oa_folder}
\title{Read a folder of ObservAir data}
\usage{
read_oa_folder(
  file_folder,
  time_zone,
  columns = c("TS", "Iref", "Isig", "ATN", "BC", "RH", "T", "FR", "Vbat", "GPSlat",
    "GPSlong")
)
}
\arguments{
\item{file_folder}{Location of data files. Files in the folder should use the following naming conventions:
\describe{
\item{Data...}{Data files should start with "Data". Example: \code{"Data34_2021-07-01.txt"}.}
\item{Settings...}{Settings file should start with "Settings". Only have one Settings file per folder. Example: \code{Settings_Data34.txt}.}
}}

\item{time_zone}{Character; valid Olson Name of time zone.
Not sure which name to use? Run \code{stringr::str_subset(OlsonNames(), "string_argument")} with \code{string_argument} replaced with an area to see names.
It is highly recommended to use a full name rather than a \code{GMT} argument.}

\item{columns}{Character list, optional; List of column names as provided in the Settings file.
If Settings file does not exist in the folder, you can list them here or use the provided default.}
}
\value{
Raw data frame containing the columns (either detected from the Settings file or specified),
with an additional \code{site_id} column for the data source.
}
\description{
Read a folder containing the resulting \code{.txt} file outputs from ObservAir
}
\examples{
\dontrun{
read_oa_folder("data-raw/oa_moving", "America/Los_Angeles")
}
}
\seealso{
Other {OA functions}: 
\code{\link{wrangle_oa_meta}()},
\code{\link{wrangle_oa}()}
}
\concept{{OA functions}}
