% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_dt_scale.R
\name{settings_dt_scale}
\alias{settings_dt_scale}
\title{Visualization Settings for a Date/Time Scale}
\usage{
settings_dt_scale(
  dataset,
  start_date = input_startdate,
  end_date = input_enddate
)
}
\arguments{
\item{dataset}{The data set which to evaluate}

\item{start_date}{The start date (to be pasted into captions); will be automatically adjusted based on actual start date of the data set, if possible}

\item{end_date}{The end date (to be pasted into captions); will be automatically adjusted based on actual end date of the data set, if possible}
}
\value{
List of values to be applied to various visualization methods.
}
\description{
A series of arguments to detect the minimum time stamp value within a data set, and provide an output of visualization settings to apply to allow for appropriate labeling and scaling.
}
\examples{
settings_results <- settings_dt_scale(data_pm25)
ggplot() + labs(subtitle = settings_results$lab_caption)
}
