% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_line.R
\name{ts_line}
\alias{ts_line}
\title{Time Series for multiple monitors' temporal atmospheric data}
\usage{
ts_line(
  dataset,
  variable_of_interest,
  add_extrema = TRUE,
  digits = 2,
  add_points = FALSE,
  add_average = TRUE,
  label_filter = c(""),
  cap_value = NA,
  cap_color = "green",
  single_column = FALSE,
  label_length = 10,
  avg_color = "darkgreen",
  min_color = "royalblue",
  max_color = "darkorange",
  location_data = data_meta
)
}
\arguments{
\item{dataset}{The hourly dataset to visualize}

\item{variable_of_interest}{The variable of interest (not in quotation marks) which to visualize}

\item{add_extrema}{Logical; label the extreme (minimum and maximum) valuess of the monitors of interest}

\item{digits}{Numeric; the number of digits to report}

\item{add_points}{Logical; add colored points for the \code{variable_of_interest}}

\item{add_average}{Logical; add a seprate plot for the average of ALL monitors in a set, to be a line colored by \link{avg_color}}

\item{label_filter}{Character list, optional (but recommended for large sets); list of monitor labels to be spotlighted (note: does NOT remove unspecified labels, which will remain as gray background lines)}

\item{cap_value}{Numeric, optional; values at or above to be colored serpately from the regular continuous scale. See \link{add_cap} for more information.}

\item{cap_color}{Character; color for values at or above the \code{cap_value}}

\item{single_column}{Logical; plot the data all as a single column, making cross-comparisons for monitors at specific times easier}

\item{label_length}{Character; applied if \code{single_column} is \code{TRUE}, the number of characters for which to wrap the monitor label}

\item{avg_color, min_color, max_color}{Character; colors for the average line, minimum text label, and maximum text label, respectively}

\item{location_data}{Data set containing latitude and longitude data}
}
\value{
Data visualization: line graph time series spotlighting individual monitors, with background graphics representing the time series data for all other monitors in the set.
}
\description{
Visualize temporal atmospheric data for multiple monitors, with optional minimum and maximum labels and a moving average.
Relevant information (such as date ranges, averaging methods, facets, and min/max values in the set) will be reported autmatically in the visualization.
}
\examples{
ts_line(july_api_hourly, pm25_atm, label_filter = "\\\\bPSU\\\\b", location_data = july_api_meta)
ts_line(july_api_hourly, pm25_atm, label_filter = c("\\\\bPSU\\\\b"), add_points = TRUE, single_column = TRUE, add_average = FALSE, location_data = july_api_meta)
}
