% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data.R
\name{wrangle_data}
\alias{wrangle_data}
\title{Wrangle PM Data}
\usage{
wrangle_data(raw_pm_data = raw_data)
}
\arguments{
\item{raw_pm_data}{The dataset which to wrangle.}
}
\value{
Dataframe with selected columns of interest.
}
\description{
Apply quality control functions (\link{apply_qc}), timezone adjustments (\link{unit_convert} and \link{adjust_timezone}), and temperature conversions (\link{ambient_temperature}) to a provided raw dataset.
}
\examples{
data_pm25 <- wrangle_data(raw_data)
}
