% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_frm.R
\name{wrangle_frm}
\alias{wrangle_frm}
\title{Wrangle federal reference monitor (FRM)}
\usage{
wrangle_frm(data_frm_raw, data_frm_meta, timezone_original = "Etc/GMT+8")
}
\arguments{
\item{data_frm_raw}{Raw FRM data set to wrangle.}

\item{data_frm_meta}{User-created location data for the FRM (key component: presence of a \code{site_id} and \code{timezone}).}

\item{timezone_original}{Character of valid Olson name; the original timezone of the raw file.
For the Oregon DEQ, it is in Pacific Standard Time, meaning DST is NOT accounted for.
Using Etc/GMT+8 here will mark the time stamps as the appropriate -8 ("PST").}
}
\value{
A data frame with quality-controlled values (\link{apply_qc}), timezone correction (\link{adjust_timezone}), and ambient temperature calculations (\link{ambient temperature}).
}
\description{
The following is based on inputs assuming an Excel input from \href{https://oraqi.deq.state.or.us/report/SingleStationReport}{Oregon DEQ single station report}.
}
\examples{
\dontrun{
2+2
}
}
\seealso{
Other {FRM functions}: 
\code{\link{read_frm}()}
}
\concept{{FRM functions}}
