% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_local.R
\name{wrangle_local}
\alias{wrangle_local}
\title{Wrangle local raw PurpleAir data frames}
\usage{
wrangle_local(data_local_raw, data_local_meta)
}
\arguments{
\item{data_local_raw}{Character; the file path for the folder that contains the monitor data frames}

\item{data_local_meta}{Character, optional; filter for a specific location (by label in the file title)}
}
\value{
A data frame with quality-controlled values (\link{apply_qc}), timezone correction (\link{adjust_timezone}), and ambient temperature calculations (\link{ambient temperature}).
}
\description{
Applying various quality control functions, time rounding, and pivots to provide a more managable data frame of spatio-temporal atmospheric data.
}
\examples{
example_local_raw <- read_local_folder(file_path = paste0(path.package("AirVizR"), "/data-raw/local"))
example_local_meta <- read_local_meta(file_path = paste0(path.package("AirVizR"), "/data-raw/local"), timezone = "America/Los_Angeles")
wrangle_local(example_local_raw, example_local_meta)
remove(example_local_raw, example_local_meta)
}
