% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_meta.R
\name{wrangle_meta}
\alias{wrangle_meta}
\title{Wrangle Location (Meta) Data}
\usage{
wrangle_meta(raw_location_data = raw_meta)
}
\arguments{
\item{raw_location_data}{The data set which to wrangle.}
}
\value{
Dataframe with selected columns of interest.
}
\description{
Select variables of interest from location data, and (importantly!) make monitor labels unique from one another.
}
\examples{
wrangle_meta(july_api_raw_meta)
}
\seealso{
\code{\link{wrangle_frm()}} and \code{\link{wrangle_local()}}

Other {miscellaneous functions}: 
\code{\link{adjust_timezone}()},
\code{\link{ambient_temperature}()},
\code{\link{apply_date_tags}()},
\code{\link{apply_hour_tags}()},
\code{\link{column_dt}()},
\code{\link{filter_df}()},
\code{\link{group_stad}()},
\code{\link{organize_stad}()},
\code{\link{rounding_w_zeroes}()},
\code{\link{unit_convert}()}
}
\concept{{miscellaneous functions}}
