% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_oa.R
\name{wrangle_oa}
\alias{wrangle_oa}
\title{Wrangle ObservAir data}
\usage{
wrangle_oa(dataset, drop_minutes = 5)
}
\arguments{
\item{dataset}{Data set for which to wrangle}

\item{drop_minutes}{Numeric; number of minutes at the start of each file to remove}
}
\value{
Wrangled data frame
}
\description{
Wrangle imported ObservAir data frames.
}
\examples{
\dontrun{
wrangle_oa(raw_oa)
}
}
\seealso{
\link{unit_convert()}

Other {OA functions}: 
\code{\link{read_oa_folder}()},
\code{\link{wrangle_oa_meta}()}
}
\concept{{OA functions}}
