% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_oa_meta.R
\name{wrangle_oa_meta}
\alias{wrangle_oa_meta}
\title{Create meta data set for OA data}
\usage{
wrangle_oa_meta(dataset, timezone, single = TRUE)
}
\arguments{
\item{dataset}{Blah}

\item{timezone}{Character}

\item{single}{Logical}
}
\value{
Data set.
}
\description{
For OA data that does not involve movement, create a meta data frame with this.
}
\examples{
2+2
}
\seealso{
\code{\link{wrangle_meta()}}

Other {OA functions}: 
\code{\link{read_oa_folder}()},
\code{\link{wrangle_oa}()}
}
\concept{{OA functions}}
