% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_timezone.R
\name{adjust_timezone}
\alias{adjust_timezone}
\title{Adjust timezone}
\usage{
adjust_timezone(dataset, location_data = raw_meta)
}
\arguments{
\item{dataset}{The dataset for which to convert timezone data (column: "datetime")}

\item{location_data}{The dataset containing information on the timezone (column: "timezone") to apply}
}
\value{
Dataset with a column of corrected timestamp ("datetime") and the original ("datetime_utc")
}
\description{
Time zones are reported as a variable when downloading the data. Time stamps are reported in UTC. The following will convert the time stamps to said reported time zone.
If more than one time zone is reported in the data, the conversion will use the time zone most frequently used in the data set. This is because no more than one time zone can be applied to a date time variable.
}
\examples{
adjust_timezone(raw_data)
}
