% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_timezone.R
\name{adjust_timezone}
\alias{adjust_timezone}
\title{Adjust timezone}
\usage{
adjust_timezone(dataset, location_data = raw_meta)
}
\arguments{
\item{dataset}{The data set for which to convert timezone data (column: "datetime")}

\item{location_data}{The data set containing information on the timezone (column: "timezone") to apply}
}
\value{
Dataset with a column of corrected time stamp ("datetime") and the original ("datetime_utc")
}
\description{
Time zones are reported as a variable when downloading the data. Time stamps in PurpleAir data are reported in UTC. The following will convert the time stamps to said reported time zone.
If more than one time zone is reported in the data, the conversion will use the time zone most frequently used in the data set. This is because no more than one time zone can be applied to a date time variable.
}
\examples{
adjust_timezone(head(july_api_raw), location_data = july_api_raw_meta)
}
\seealso{
Other {miscellaneous functions}: 
\code{\link{ambient_temperature}()},
\code{\link{apply_date_tags}()},
\code{\link{apply_hour_tags}()},
\code{\link{column_dt}()},
\code{\link{filter_df}()},
\code{\link{group_stad}()},
\code{\link{organize_stad}()},
\code{\link{rounding_w_zeroes}()},
\code{\link{unit_convert}()},
\code{\link{wrangle_meta}()}
}
\concept{{miscellaneous functions}}
