% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirVizR-package.R
\docType{package}
\name{AirVizR-package}
\alias{AirVizR}
\alias{AirVizR-package}
\title{AirVizR: Visualization Tools for Air Quality Data}
\description{
A series of tools for extracting, wrangling, correcting, and visualizing air quality data, with an emphasis on PM2.5 data from PurpleAir.
}
\author{
\strong{Maintainer}: Gillian McGinnis \email{gmcginnis@reed.edu}

}
\keyword{internal}
