% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient_temperature.R
\name{ambient_temperature}
\alias{ambient_temperature}
\title{Ambient Temperature}
\usage{
ambient_temperature(
  dataset,
  variable = temperature,
  change = temperature_change
)
}
\arguments{
\item{dataset}{The dataset containing temperature values to adjust}

\item{variable}{The column name of temperature data (in ºF) which to apply the change}

\item{change}{Numeric: The value which to subtract from the original temperature value}
}
\value{
Dataset with two new columns for ambient temperature:
\describe{
\item{temperature_ambient}{Ambient temperature in ºF}
\item{temperature_ambient_c}{Ambient temperature in ºC}
}
}
\description{
PurpleAir monitor reports include internal temperature, which is on average slightly higher than ambient temperature.
More information can be found on \href{https://www2.purpleair.com/community/faq#!hc-primary-and-secondary-data-header}{their FAQ}.
}
\examples{
july_api_raw \%>\% 
  ungroup() \%>\% 
  slice(1:10) \%>\% 
  ambient_temperature(change = 8)
}
