% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient_temperature.R
\name{ambient_temperature}
\alias{ambient_temperature}
\title{Ambient Temperature}
\usage{
ambient_temperature(dataset, variable = temperature, change = 8)
}
\arguments{
\item{dataset}{The dataset containing temperature values to adjust}

\item{variable}{The column name of temperature data (in ºF) which to apply the change}

\item{change}{Numeric: The value which to subtract from the original temperature value}
}
\value{
Dataset with two new columns for ambient temperature, and temperature in ºC:
\describe{
\item{temperature}{Ambient temperature in ºF}
\item{temperature_c}{Ambient temperature in ºC}
\item{temperature_inernal}{Internal temperature in ºF}
\item{temperature_internal_c}{Internal temperature in ºC}
}
}
\description{
PurpleAir monitor reports include internal temperature, which is on average slightly higher than ambient temperature.
More information can be found on \href{https://www2.purpleair.com/community/faq#!hc-primary-and-secondary-data-header}{their FAQ}.
}
\examples{
ambient_temperature(head(july_api_raw[1:3]))
}
\seealso{
\code{\link{unit_convert()}}

Other {miscellaneous functions}: 
\code{\link{adjust_timezone}()},
\code{\link{apply_date_tags}()},
\code{\link{apply_hour_tags}()},
\code{\link{column_dt}()},
\code{\link{filter_df}()},
\code{\link{group_stad}()},
\code{\link{organize_stad}()},
\code{\link{rounding_w_zeroes}()},
\code{\link{unit_convert}()},
\code{\link{wrangle_meta}()}
}
\concept{{miscellaneous functions}}
