% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_hour_tags.R
\name{apply_hour_tags}
\alias{apply_hour_tags}
\title{Create a column of "Date Tags"}
\usage{
apply_hour_tags(dataset, starts = input_hour_starts, tags = input_hour_tags)
}
\arguments{
\item{dataset}{The dataset for which to apply the date tags to.}

\item{starts}{Numeric list of start hours (in 24 hour format)}

\item{tags}{Character list of tags}
}
\value{
Dataframe with a new column ("hour_tag") containing the appropriate hour tag with respect to each row's timestamp.
}
\description{
Create a column of date tags based on specified hour ranges (created using inputted \code{starts}) and respective names (\code{tags}).
}
\examples{
apply_hour_tags(
  july_frm_diurnal[1:3],
  c(5, 12, 17, 21),
  c("Morning", "Afternoon", "Evening", "Night")
)
apply_hour_tags(
  july_frm_diurnal[1:3],
  starts = c(6, 22),
  tags = c("Awake", "Asleep")
)
}
\seealso{
Other {miscellaneous functions}: 
\code{\link{adjust_timezone}()},
\code{\link{ambient_temperature}()},
\code{\link{apply_date_tags}()},
\code{\link{column_dt}()},
\code{\link{filter_df}()},
\code{\link{group_stad}()},
\code{\link{organize_stad}()},
\code{\link{rounding_w_zeroes}()},
\code{\link{unit_convert}()},
\code{\link{wrangle_meta}()}
}
\concept{{miscellaneous functions}}
