% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_lrapa.R
\name{apply_lrapa}
\alias{apply_lrapa}
\title{Apply LRAPA Corrections}
\source{
\url{https://www.lrapa.org/DocumentCenter/View/4147/PurpleAir-Correction-Summary}

\href{https://www.lrapa.org/DocumentCenter/View/4147/PurpleAir-Correction-Summary}{LRAPA documentation}
}
\usage{
apply_lrapa(dataset, by_day = TRUE, by_hour = FALSE, keep_cols = FALSE)
}
\arguments{
\item{dataset}{The dataset for which to apply the correction factors to}

\item{by_day}{Logical; average data by day}

\item{by_hour}{Logical; average data by hour}

\item{keep_cols}{Logical; Keep or disgard extra columns. If FALSE, only identifying columns and LRAPA-corrected columns will remain}
}
\value{
Dataframe with new column for LRAPA-corrected PM2.5:
\describe{
\item{pm25_lrapa}{LRAPA-corrected PM2.5 value, calculated for values where PM2.5(CF=1) ≤ 65 µg/m^3 as 0.5 × PM2.5(CF=ATM) - 0.66}
}
}
\description{
Apply LRAPA correction factors to PurpleAir PM2.5 data.
\href{https://www.lrapa.org/}{LRAPA} established PM2.5 correction factors for PurpleAir's PM2.5 CF=ATM values.
Unlike the EPA correction factors (\link{apply_epa}) which uses CF=1 PM2.5 values, the LRAPA correction factor uses CF=ATM values. More information on this variable can be found on the \href{https://www2.purpleair.com/community/faq#hc-what-is-the-difference-between-cf-1-and-cf-atm}{PurpleAir FAQ}.
It should also be noted that this correction factor might not apply to all airsheds, and was developed in Oregon, where (along with much of the Pacific Northwest) woodsmoke is a major contributor to PM2.5.
Futhermore, this correction factor only reliably applies to PM2.5 values of ≤ 65 µg/m^3.
For more information on the correction factor, see the \href{https://www.lrapa.org/DocumentCenter/View/4147/PurpleAir-Correction-Summary}{LRAPA documentation}.
}
\examples{
apply_lrapa(july_api_full)
\donttest{apply_lrapa(july_api_full, by_hour = TRUE, keep_cols = TRUE)}
}
\seealso{
Other {PA functions}: 
\code{\link{apply_corrections}()},
\code{\link{apply_epa}()},
\code{\link{apply_functions}()},
\code{\link{apply_qc}()}
}
\concept{{PA functions}}
