% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_qc.R
\name{apply_qc}
\alias{apply_qc}
\title{Apply Quality Control Filters}
\usage{
apply_qc(dataset, drop_hi = input_drop_hi, avg_ab = TRUE, loc_data = raw_meta)
}
\arguments{
\item{dataset}{The dataset for which to convert timezone data (column: "datetime")}

\item{drop_hi}{Logical; drop monitors for which "flag_highValue" is TRUE in the defined meta data}

\item{avg_ab}{Logical; create a column to average PM2.5 values of A & B monitors}

\item{loc_data}{The dataset containing site IDs and flag_highValue}
}
\value{
Dataset with improbable values removed, and (optionally) additional columns for averaged A & B monitor values:
\describe{
\item{pm25_cf1}{Rowwise mean of \code{pm25_cf1_A} and \code{pm25_cf1_B}}
\item{pm25_atm}{Rowwise mean of \code{pm25_atm_A} and \code{pm25_atm_B}}
}
}
\description{
Apply basic quality control to a dataset: PM2.5 between 0 and 2000, temperature (ºF) between -40 and 185, and humidity (\%) between 0 and 100.
Also provides an option to average PM2.5 columns A and B of a dataset
}
\examples{
\donttest{
apply_qc(july_api_raw, drop_hi = TRUE, loc_data = july_api_raw_meta)
}
}
