% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_dt.R
\name{column_dt}
\alias{column_dt}
\title{Adding new date/time column(s)}
\usage{
column_dt(dataset, unit)
}
\arguments{
\item{dataset}{The dataset for which to convert timezone data (column: "datetime")}

\item{unit}{The unit(s) for which to apply to the dataset. Options: "date", "date_hour", "hour", "hour_minute", "time"}
}
\value{
Dataset with column(s) of newly specified time unit
}
\description{
Creating a new column in a data set with a specified date or time rounding/conversion
}
\examples{
column_dt(raw_data, c("date", "hour"))
}
