% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_incomplete.R
\name{drop_incomplete}
\alias{drop_incomplete}
\title{Drop Incomplete Sets}
\usage{
drop_incomplete(dataset, var_qt)
}
\arguments{
\item{dataset}{Data set for which to filter incomplete monitor data}

\item{var_qt}{Character; the variable of interest (in quotation marks) for which to apply the cap}
}
\value{
Data set with monitors without complete sets removed
}
\description{
Drop monitor data for which its time series data is incomplete compared to that of other monitors in the data set.
}
\examples{
drop_incomplete(july_api_diurnal, "pm25_atm") \%>\% ggplot(aes(hour, site_id, fill = pm25_atm)) + geom_tile()
}
