% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_pas.R
\name{get_area_pas}
\alias{get_area_pas}
\title{Getting PurpleAir Sensors (PAS) by Area and Date}
\usage{
get_area_pas(
  state_code = input_stateCode,
  west = input_west,
  east = input_east,
  south = input_south,
  north = input_north,
  labels = input_labels,
  datestamp = input_enddate,
  startdate = input_startdate
)
}
\arguments{
\item{state_code}{Character, optional. State of monitors to include. If NULL/unspecified, all monitors specified in the above boundaries will be included regardless of state.}

\item{west}{Numeric; The western boundary coordinate}

\item{east}{Numeric; The eastern boundary coordinate}

\item{south}{Numeric; The southern boundary coordinate}

\item{north}{Numeric; The northern boundary coordinate}

\item{labels}{Character list, optional; Labels of monitors to include. If \code{NULL}/unspecified, all monitors will be included. The labels are based on string detection (\link[stringr]{str_detect}), so for instance having "STAR" will pull all that have the word "STAR" in the label. Capitalization matters.}

\item{datestamp}{Character; Important argument for loading historical data, as not all monitors might be actively reporting anymore. Format: "YYYY-MM-DD".}

\item{startdate}{Character; Relevant for calculating lookback days (the maximum number of days to go back and try to load data if the requested date cannot be retrieved). Format: "YYYY-MM-DD".}
}
\value{
A dataframe returning all PAS within the defined area
}
\description{
Load the PAS within a specified area
}
\examples{
get_area_pas(
  "OR",
  -122.854, 45.4, -122.58, 45.6,
  c("se", "SE", "Se", "\\\\bSTAR\\\\b", "\\\\bPSU\\\\b"),
  datestamp = "2020-07-07", startdate = "2020-07-01"
)
}
