% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_pat.R
\name{get_area_pat}
\alias{get_area_pat}
\title{Get raw timeseries and sensor meta data}
\usage{
get_area_pat(
  id_list = ids,
  startdate = input_startdate,
  enddate = input_enddate,
  cols = c("created_at", "temperature", "humidity", "pm2.5_cf1", "pm2.5_atm")
)
}
\arguments{
\item{id_list}{A list of PAS IDs}

\item{startdate}{Date at which to start the collection (format: "YYYY-MM-DD")}

\item{enddate}{Date through which the data will be collected (format: "YYYY-MM-DD")}

\item{cols}{List of variables of interest to keep. Site ID info will autmatically be collected}
}
\value{
List: dataframes of sensor meta info, and a dataframe of raw data for the variables of interest
}
\description{
Iterate over a list of PurpleAir sensor IDs to extract sensor meta data and timeseries data given specified start and end dates
}
\examples{
results <- get_area_pat(startdate = "2021-07-01", enddate = "2021-07-07")
meta_data <- results$raw_meta
raw_data <- results$raw_data
}
