% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_ids}
\alias{get_ids}
\title{Get IDs from a PAS}
\usage{
get_ids(pas = pas_area, outside = include_outside, inside = include_inside)
}
\arguments{
\item{pas}{A dataframe of PAS containing a column of device deployment IDs}

\item{outside}{Logical argument to include/exclude outdoor monitors}

\item{inside}{Logical argument to include/exclude indoor monitors}
}
\value{
A list of IDs from a provided PAS
}
\description{
Load a list of device deployment IDs from a provided PAS
}
\examples{
get_ids((example(get_area_pas))$value, TRUE, TRUE)
\donttest{
get_ids(
  pas = AirSensor::pas_filter(AirSensor::pas_load(datestamp = "20200701"), stateCode == "OR"),
  outside = TRUE, inside = TRUE
)
}
}
\seealso{
\code{\link[AirSensor]{pas_getDeviceDeploymentIDs()}}

Other {PA API functions}: 
\code{\link{get_area_pas}()},
\code{\link{get_area_pat}()},
\code{\link{wrangle_data}()}
}
\concept{{PA API functions}}
