% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_stad.R
\name{group_stad}
\alias{group_stad}
\title{Average Data by Time Groupings}
\usage{
group_stad(dataset, by_day = TRUE, by_hour = FALSE)
}
\arguments{
\item{dataset}{The dataset for which to average}

\item{by_day}{Logical; average data by day}

\item{by_hour}{Logical; average data by hour}
}
\value{
Dataframe with numeric values averaged by specified timestamp(s)
}
\description{
Average numeric values in a dataset by day and/or by hour.
}
\examples{
data_daily <- group_stad(july_api_full, by_day = TRUE)
data_diurnal <- group_stad(july_api_full, by_day = FALSE, by_hour = TRUE)
}
