% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stad.R
\name{map_stad}
\alias{map_stad}
\title{Map Spatio-Temporal Atmospheric Data (STAD)}
\usage{
map_stad(
  dataset,
  variable_of_interest,
  grouping_vars = NULL,
  location_data = data_meta,
  cap_value = NA,
  cap_color = "red",
  point_size = 3,
  maptype = "toner-lite",
  zoom = 11,
  tint_alpha = 0.5,
  tint_color = "black"
)
}
\arguments{
\item{dataset}{Data set for which to visualize}

\item{variable_of_interest}{The variable of interest (not in quotation marks) which to visualize}

\item{grouping_vars}{Character, optional; one or two variables for which to facet (grid) the plot by}

\item{location_data}{Data set containing latitude and longitude data}

\item{cap_value}{Numeric, optional; values at or above to be colored serpately from the regular continuous scale. See \link{add_cap} for more information.}

\item{cap_color}{Character; color for values at or above the \code{cap_value}}

\item{point_size}{Numeric; size of the points on the map}

\item{maptype}{Character; the \link[ggmap]{qmplot} class of maptype to use; see \link[ggmap]{get_map} for types.}

\item{zoom}{Numeric; the zoom level of the map}

\item{tint_alpha}{Numeric; the transparency for the background tint overlaid on the map}

\item{tint_color}{Character; the color of background tint overlaid on the map}
}
\value{
Data visualization: map with data points colored by a specified numeric variable, located by a provided data frame of lat/long data.
}
\description{
Visualize spatio-temporal atmospheric data using points on a map, colored by a specified variable.
Relevant information (such as date ranges, averaging methods, facets, and min/max values in the set) will be reported automatically in the visualization.
}
\examples{
\donttest{
map_stad(july_api_daily, pm25_atm, location_data = july_api_meta, grouping_vars = "date_tag")
map_stad(
  july_api_daily, pm25_atm, location_data = july_api_meta, grouping_vars = "date_tag",
  point_size = 5, cap_value = 50, cap_color = "green",
  maptype = "terrain", tint_color = "white"
)
}
}
