% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_local_meta.R
\name{read_local_meta}
\alias{read_local_meta}
\title{Create a data frame of locally loaded monitor data}
\usage{
read_local_meta(
  file_path = input_path,
  pattern = NULL,
  timezone = input_timezone
)
}
\arguments{
\item{file_path}{Character; the file path for the folder that contains the monitor data frames}

\item{pattern}{Character, optional; filter for a specific location (by label in the file title)}

\item{timezone}{Character of valid Olson name for timezone of the monitors}
}
\value{
A wrangled data frame of monitor location data from a local folder of PurpleAir data
}
\description{
Load PurpleAir monitor data based on locally stored data, extracting relevant information from the file names. See
}
\examples{
read_local_meta(file_path = paste0(path.package("AirVizR"), "/data-raw/local"), timezone = "America/Los_Angeles")
read_local_meta(file_path = paste0(path.package("AirVizR"), "/data-raw/local"), pattern = "Lincoln", timezone = "America/Los_Angeles")
}
