% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_w_zeroes.R
\name{rounding_w_zeroes}
\alias{rounding_w_zeroes}
\title{Round a Value, Keeping Trailing Zeroes}
\usage{
rounding_w_zeroes(num, input_digits)
}
\arguments{
\item{num}{Numeric; value(s) to round}

\item{input_digits}{Numeric; the number of digits which to preserve}
}
\value{
Character versions rounded to the number of digits specified, including trailing zeroes
}
\description{
When pasting labels, R will not display trailing zeroes; this is a way to work around it. The result can be returned to a numeric class if desired.
It should be noted that if the digit input exceeds the actual digits in the value, it will still paste a trailing zero.
}
\examples{
rounding_w_zeroes(10.501, 2)
rounding_w_zeroes(10.591, 2)
}
\seealso{
Other {miscellaneous functions}: 
\code{\link{adjust_timezone}()},
\code{\link{ambient_temperature}()},
\code{\link{apply_date_tags}()},
\code{\link{apply_hour_tags}()},
\code{\link{column_dt}()},
\code{\link{filter_df}()},
\code{\link{group_stad}()},
\code{\link{organize_stad}()},
\code{\link{unit_convert}()},
\code{\link{wrangle_meta}()}
}
\concept{{miscellaneous functions}}
