% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_dt_scale.R
\name{settings_dt_scale}
\alias{settings_dt_scale}
\title{Visualization Settings for a Date/Time Scale}
\usage{
settings_dt_scale(
  dataset,
  start_date = input_startdate,
  end_date = input_enddate
)
}
\arguments{
\item{dataset}{The data set which to evaluate}

\item{start_date}{The start date (to be pasted into captions); will be automatically adjusted based on actual start date of the data set, if possible}

\item{end_date}{The end date (to be pasted into captions); will be automatically adjusted based on actual end date of the data set, if possible}
}
\value{
List of values to be applied to various visualization methods:
\describe{
\item{dataset}{Dataset with the minimum time stamp column header renamed 'timestamp'.}
\item{lab_title_sub}{Character; Working in conjunction with \code{lab_title} from \link{settings_units}, the unit of temporal measurement of the dataset}
\item{x_angle}{Numeric; the angle to place x-axis (temporal) text.}
\item{x_scale}{x-axis scale (from \link[ggplot2]{scale_x_datetime}) formatting and scale breaks.}
\item{date_in_set}{Logical; if date value(s) present in data set.}
\item{start_date}{Date; actual minimum date of data set, regardless of function input.}
\item{end_date}{Date; actual maximum date of data set, regardless of function input.}
\item{lab_caption}{Character; label information for details containing date range in the set.}
}
}
\description{
A series of arguments to detect the minimum time stamp value within a data set, and provide an output of visualization settings to apply to allow for appropriate labeling and scaling.
}
\examples{
\donttest{
settings_results_example <- settings_dt_scale(july_api_diurnal, start_date = "2020-07-01", end_date = "2020-07-07")
ggplot(settings_results_example$dataset, aes(timestamp, temperature)) +
  settings_results_example$x_scale +
  labs(subtitle = settings_results_example$lab_caption)
remove(settings_results_example)
}
}
\seealso{
Other {visualization settings}: 
\code{\link{add_cap}()},
\code{\link{drop_incomplete}()},
\code{\link{settings_shapes}()},
\code{\link{settings_units}()}
}
\concept{{visualization settings}}
