% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_shapes.R
\name{settings_shapes}
\alias{settings_shapes}
\title{Visualization Settings for Monitor Locations}
\usage{
settings_shapes(dataset = dataset, meta = data_meta)
}
\arguments{
\item{dataset}{The dataset which to evaluate}

\item{meta}{The location data, related the data set}
}
\value{
List of values to be applyed to various visualization methods.
}
\description{
Arguments to detect the relevant shapes to be applied to visualizations based on monitor location(s) in the data set.
Custom shapes are necessary for the custom visualizations, since "fill" and "color" arguments can be individually specified for the shapes below, unlike the default shapes.
}
\examples{
shape_results <- settings_shapes(data_pm25, data_meta)
ggplot() + shape_set + shape_guide
}
