% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_shapes.R
\name{settings_shapes}
\alias{settings_shapes}
\title{Visualization Settings for Monitor Locations}
\usage{
settings_shapes(dataset = dataset, meta = data_meta)
}
\arguments{
\item{dataset}{The dataset which to evaluate}

\item{meta}{The location data, related the data set}
}
\value{
List of values to be applied to various visualization methods:
\describe{
\item{shape_set}{Scale settings for shapes; see \link[ggplot2]{scale_shape_manual}.}
\item{shape_guide}{Guide settings for shapes; see \link[ggplot2]{guides}.}
}
}
\description{
Arguments to detect the relevant shapes to be applied to visualizations based on monitor location(s) in the data set.
Custom shapes are necessary for the custom visualizations, since "fill" and "color" arguments can be individually specified for the shapes below, unlike the default shapes.
}
\examples{
shape_results <- settings_shapes(july_api_daily, july_api_meta)
ggplot(full_join(july_api_daily, july_api_meta), aes(date, site_id, shape = location)) + shape_results$shape_set + shape_results$shape_guide + geom_point()
}
