% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_units.R
\name{settings_units}
\alias{settings_units}
\title{Visualization Settings for Specified Units}
\usage{
settings_units(
  dataset = dataset,
  var_qt = variable_of_interest_qt,
  cap_value = NA,
  cap_color = "red",
  digits = 2,
  lab_title = "Graph of",
  lab_fill = "Units",
  lab_unit = "units"
)
}
\arguments{
\item{dataset}{The dataset which to evaluate}

\item{var_qt}{Character; The variable of interest within the dataset, in quotation marks}

\item{cap_value}{Numeric; The value of the color "cap" to be applied}

\item{cap_color}{Character; The color of values at or above the \code{cap_value}.}

\item{digits}{Numeric; the number of digits to round data labels to.}

\item{lab_title}{Character; the first portion of the visualization title. It is recommended to adjust this to the type of visualization to be made.}

\item{lab_fill}{Character; the default unit label for the variable of interest (standalone, e.g. y-axis label of 'Units')}

\item{lab_unit}{Character; the default unit label for the variable of interest (in context of a sentence, e.g. 'X units')}
}
\value{
List of values to be applyed to various visualization methods.
}
\description{
A series of arguments to detect the appropriate visualization labels and settings to apply to a dataset provided said dataset and a varaible of interest.
Additionally, an optional "cap" value (and custom color) can be provided in order to distinguish values at or above said specified values to prevent color scale washout (e.g. if a few observations are extremely high to the point where all other values are washed out on the color scale).
}
\examples{
units_results <- settings_units(data_pm25, "pm25_atm", cap_value = 150)
ggplot() + fill_colors + cap_guide + labs(fill = units_results$lab_fill)
}
