% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_convert.R
\name{unit_convert}
\alias{unit_convert}
\title{Convert Temperature}
\usage{
unit_convert(fahrenheit, celsius)
}
\arguments{
\item{fahrenheit}{Numeric: A value of temperature in ºF to be converted to ºC}

\item{celsius}{Numeric: A value of temperature in ºC to be converted to ºF}
}
\value{
Value in the other unit
}
\description{
Convert temperature from Fahrenheit to Celsius, or vice versa.
}
\examples{
unit_convert(fahrenheit = 23)
unit_convert(celsius = 23)
dplyr::mutate(head(july_api_raw), temperature_c = unit_convert(fahrenheit = temperature))
}
