% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_meta.R
\name{wrangle_meta}
\alias{wrangle_meta}
\title{Wrangle Location (Meta) Data}
\usage{
wrangle_meta(raw_location_data = raw_meta)
}
\arguments{
\item{raw_location_data}{The dataset which to wrangle.}
}
\value{
Dataframe with selected columns of interest.
}
\description{
Select variables of interest from location data, and (importantly) make monitor labels unique from one another.
}
\examples{
data_meta <- wrangle_meta(july_api_raw_meta)
}
