% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_oa_meta.R
\name{wrangle_oa_meta}
\alias{wrangle_oa_meta}
\title{Create meta data set for OA data}
\usage{
wrangle_oa_meta(dataset, timezone, single = TRUE)
}
\arguments{
\item{dataset}{Wrangled OA data set}

\item{timezone}{Character of valid Olson Name}

\item{single}{Logical; group all the data into a singular "location", or keep seprate?}
}
\value{
Data set.
}
\description{
For OA data that does not involve movement, create a meta data frame with this.
}
\examples{
wrangle_oa_meta(oa_moving_full, "America/Los_Angeles")
wrangle_oa_meta(oa_static_full, "America/Los_Angeles", TRUE)
}
\seealso{
Other {OA functions}: 
\code{\link{read_oa_folder}()},
\code{\link{wrangle_oa}()}
}
\concept{{OA functions}}
